change repository key to Integer and added a findOne method in bookservice/repo

+++ project_4_web_service/src/main/java/io/ab/library/repository/AuthorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import io.ab.library.model.Author;
 
-public interface AuthorRepository extends CrudRepository<Author, Long> {
+public interface AuthorRepository extends CrudRepository<Author, Integer> {
 
 }
